Add service tests for contact internal route. Fix middleware issue. Split last contact service tests into specific files
